Move sample config func to separate file
