+++ plugins/aggregators/starlark/starlark_sample_config.go
@@ -1,0 +1,6 @@
+// DON'T EDIT; This file is used as a template by tools/generate_plugindata
+package starlark
+
+func (s *Starlark) SampleConfig() string {
+	return `{{ .SampleConfig }}`
+}
